print_command_error to lib
